LayoutTests/inspector/css/modify-css-property-expected.txt

Testing that CSSStyleDeclaration update immediately after modifying its properties when it is not locked.


== Running test suite: ModifyCSSProperty
-- Running test case: Update value when CSSStyleDeclaration is not locked
PASS: "font-size" property value should update immediately.
PASS: Style declaration text should stay unchanged.

-- Running test case: Update value when CSSStyleDeclaration is locked
PASS: "font-size" property value should update immediately.
PASS: Style declaration text should update immediately.

-- Running test case: Update inline style value when CSSStyleDeclaration locked and not locked
PASS: Style declaration text should update immediately.
PASS: Style declaration text should stay "width: 64px".
PASS: "width" property value should update to "200px".
PASS: Inline style declaration text should update when not locked.
